Make DefaultFeeOracle and DataUnionFactory upgradeable (#339)
* feat: DefaultFeeOracle and DataUnionFactory are now upgradeable

* build: add hardhat-upgrades and contracts-upgradeable as deps

* fix: remove prepare step

it's super annoying, and now it's breaking CI. No idea why it was there in the first place.

* fix: lint

* build: add missing devDependency

* test: fix deployment of upgradeable proxies

* feat: upgrade openzeppelin dependency

* fix: upgradeable init parent calls

* test: fix upgradeable initialize

* fix: attempt to fix makefile

it's probably way too long and complex to fix. This fix is clear: contracts were moved from deps to devDeps. But it's too complex to patch in the contracts-upgradeable, too, I'm afraid...

* build: remove solidity makefile build step

it's not used anywhere, and it's blocking the CI. I'll make a ticket of fixing it.

* fix: tabs to spaces

* refactor: more readable variable names

* build: harmonize hardhat and @streamr/config network names

* test: better naming

* feat: ETH-358 deployment script + artifacts

artifacts are needed for upgrading the proxy implementations later