Polygon issues debugging (#329)
* refactor: remove ternary for better readability

* feat: moved gasPriceStrategy into its own file, now it's more than just number->number mapping

this is because Polygon might (note: MIGHT) require asking their official HTTP API what the gas price is supposed to be. Also we should modify not just gasPrice to coax ethers to play nice with Polygon but all post-London fee parameters as well. This is the first iteration and I can probably drop the HTTP API later.