ETH-381: Index RevenueReceived events in DU subgraph (#441)
* feat: added RevenueEvent

* fix: add RevenueReceived handler to yaml

* fix: add default value to argument so that it works like the other deploy method

* test: add waits to setup

this way it works also nicely with dev-docker environment and not just ganache (I tested using it in thegraph-subgraph test, and it works with these changes, yay)

* refactor: remove src/log.ts, just use debug directly

* release(client): 3.0.6

* doc: fix case

* release(client): 3.0.7

uses the newest @streamr/config

* test: typo fix

* doc: add todo comment

* feat: removed deprecated event

this was used in DU2 version of thegraph-subgraph, probably that's why it had been kept. But now subgraphs don't use it anymore. It's available in production, but shouldn't be used. And it will be gone in a future update from production, too.

* feat: RevenueReceived event handling added to thegraph-subgraph, also a test

WIP: get it to run nicely in the CI. Already runs on my machine.

* build: run thegraph test in CI

* build: fix yaml lint error

* test: remove deprecated event name

* test: add build step to integration test

* refactor: clean up the AssemblyScript file

* test: deploy the subgraph before test

* build: copy ABIs from contracts to subgraph package

this should be done when the contracts are deployed to production (so that the subgraph looks like what's in production)

* feat: support old DU factories (Swash) in the subgraph

* doc: comment different create DU handlers

* build: force also old DU contract code generated in "npm run build"

only gnosis has those DataUnionSidechain and DataUnionFactorySidechain