feat: ETH-400 track DU metadata field in thegraph subgraph
left the DataUnionFactory change as TODO since, well, if it's not a problem then don't change it. But it might be a problem that the events come in wrong order, and that TODO might be the fix. Basically thegraph might need to first get DUCreated to start listening to the DU, and after that it can get the MetadataChanged from that DU. They both will happen within the same tx, so even that change might not help; in which case DataUnionFactory would need to be changed to emit the initial metadata in a separate event.